requirements.txt
# Ficheiro requirements.txt
discord.py
requests
beautifulsoup4
python-dateutil
python-dotenv